+++ src/pages/OrderPage.tsx
@@ -15,8 +15,8 @@ import { orderSchema, type OrderFormData } from '@/utils/validateOrderSchema';
 import * as z from 'zod';
 import { toast } from 'react-toastify';
 import { AxiosError, HttpStatusCode } from 'axios';
-import { useAuth } from '@/hooks/service/authHooks';
-import { useOrderMutation, useProductSummaryQuery } from '@/hooks/queries/orderHooks';
+import { useAuth } from '@/hooks/useAuthHooks';
+import { useOrderMutation, useProductSummaryQuery } from '@/queries/useOrderHooks';
 import { Suspense } from 'react';
 import Spinner from '@/components/Spinner';
 import { ErrorBoundary } from 'react-error-boundary';
